Use row, column, and square references to test for failure
Created a reference object for the rows, columns, and squares that
shows all the cells the in each applicable row, column, and square.
This allows for one failure checking function (instead of one function
for each row, column and square) that goes through the referenced
cells to look for repeated values.